# config/openapi_schema_snapshot.json
{
...
  "openapi": "3.1.0",
...
  "paths": {
...
    "/api/v1/assistant/feedback": {
      "post": {
        "deprecated": false,
        "operationId": "submit_feedback_api_v1_assistant_feedback_post",
        "responses": [
          "200",
          "422"
        ]
      }
    },
    "/api/v1/assistant/query": {
      "post": {
        "deprecated": false,
        "operationId": "query_assistant_api_v1_assistant_query_post",
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
    "/api/v1/assistant/suggest": {
      "get": {
        "deprecated": false,
        "operationId": "get_suggestions_api_v1_assistant_suggest_get",
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
    "/api/v1/feedback/": {
      "post": {
        "deprecated": false,
        "operationId": "submit_feedback_api_v1_feedback__post",
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
    "/api/v1/ocr/extract": {
      "post": {
        "deprecated": false,
        "operationId": "ocr_extract_api_v1_ocr_extract_post",
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
    "/api/v1/vision/analyze": {
      "post": {
        "deprecated": false,
        "operationId": "analyze_vision_api_v1_vision_analyze_post",
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
  },
...
}